Create examples folder Fixes #75
