+++ examples/common_lib/arm64-v8a/libVkLayer_khronos_validation.so
