+++ examples/simple-scene/assets/damaged_helmet.glb
