+++ examples/simple-scene/logo.png
