+++ test_assets/asteroid.glb
