+++ test_assets/left_hand.glb
