+++ test_assets/right_hand.glb
